+++ examples/devenv-module-recipe/devenv.nix
@@ -14,7 +14,7 @@
       install.enable = true;
 
       config = {
-        profiles = {
+        profiles.default = {
           settings.compiler."compiler.cppstd" = "17";
           settings._.build_type = "Release";
         };