+++ test/inner-home-directory/flake.nix
@@ -30,7 +30,7 @@
         {
           conan = {
             homeDirectory = "./project";
-            profiles.settings = {
+            profiles.default.settings = {
               compiler = {
                 "compiler.cppstd" = "23";
               };