+++ test/standalone-submodule-with/flake.nix
@@ -46,7 +46,7 @@
                   imports = [ conanModule ];
 
                   conan = {
-                    profiles = {
+                    profiles.default = {
                       settings.compiler."compiler.cppstd" = "14";
                       settings._.build_type = "Debug";
                     };