+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Revision history for conan-flake
 
-## 0.8.2 (unreleased)
+## 0.9.0 (unreleased)
 
 ### Improvements